Changed strcmp() return type from int to int8_t

+++ include/string.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-extern int strcmp(const char *str1, const char *str2);
+extern int8_t strcmp(const char *str1, const char *str2);
 extern size_t strlen(const char* str);
 
 #endif
